MSH Project/bmi/bank-muamalat-monitor/requirements.txt
langchain==0.1.6
openai==1.12.0
beautifulsoup4>=4.9.0
PyPDF2>=2.0.0
pdfplumber>=0.7.0
openpyxl>=3.0.0
lxml>=4.6.0
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-community